requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.71.0
async-timeout==4.0.3
bert-score==0.3.13
cachetools==6.2.1
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.1.1
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
evaluate==0.4.6
faiss-cpu==1.12.0
fastapi==0.110.0
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.8.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
hf-xet==1.1.10
httptools==0.7.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
kiwisolver==1.4.9
langchain==1.0.0
langchain-anthropic==1.0.0
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-openai==1.0.0
langchain-tavily==0.2.12
langgraph==1.0.1
langgraph-checkpoint==2.1.2
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
marshmallow==3.26.1
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
openai==2.5.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==24.2
pandas==2.3.3
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyparsing==3.2.5
python-dotenv==1.1.1
regex==2025.9.18
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
sentencepiece==0.2.1
SQLAlchemy==2.0.44
sqlite-vec==0.1.6
starlette==0.36.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.20.1
zstandard==0.25.0
